--- a/include/onnxruntime/core/framework/run_options.h
+++ b/include/onnxruntime/core/framework/run_options.h
@@ -27,6 +27,10 @@ struct OrtRunOptions {
   // So it is possible that only some of the nodes are executed.
   bool only_execute_path_to_fetches = false;
 
+  // Set to 'true' to synchronize execution providers with CPU at the end of session run.
+  // Taking CUDA EP as an example, it will trigger cudaStreamSynchronize on the compute stream.
+  bool synchronize_execution_providers = true;
+
 #ifdef ENABLE_TRAINING
   // Set to 'true' to run in training mode.
   bool training_mode = true;