--- a/setup.py
+++ b/setup.py
@@ -479,6 +479,7 @@ if enable_training:
             "onnxruntime.training.experimental",
             "onnxruntime.training.experimental.gradient_graph",
             "onnxruntime.training.optim",
+            "onnxruntime.training.torchdynamo",
             "onnxruntime.training.ortmodule",
             "onnxruntime.training.ortmodule.experimental",
             "onnxruntime.training.ortmodule.experimental.json_config",