@souptc 2022-11-01
please remind me once it is in, i need to update the multi-stream pr to apply the new option "run_options.synchronize_execution_providers"